backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
alembic
python-dotenv
pydantic
